pyproject.toml
[project]
name = "homebench"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "accelerate>=1.11.0",
    "langchain>=1.0.7",
    "langchain-ollama>=1.0.0",
    "torch>=2.9.1",
    "tqdm>=4.67.1",
    "transformers>=4.57.1",
]

[tool.uv.sources]
torch = [
  { index = "pytorch-cu128", marker = "sys_platform != 'linux'" },
]
torchvision = [
  { index = "pytorch-cu128", marker = "sys_platform != 'linux'" },
]

[tool.uv]
index-url = "https://pypi.tuna.tsinghua.edu.cn/simple"


[[tool.uv.index]]
name = "pytorch-cu128"
url = "https://download.pytorch.org/whl/cu128"
explicit = true
